refactor: DESKTOP_ONE/TWO breakpoint 공용화
- 멤버 탭에서만 쓰이던 브레이크포인트가 메인 페이지에도 쓰여서 mediaQuery로 이동
