--- apps/playground/src/components/members/main/contants.ts
@@ -1,2 +1,0 @@
-export const DESKTOP_ONE_MEDIA_QUERY = 'screen and (max-width: 1542px)';
-export const DESKTOP_TWO_MEDIA_QUERY = 'screen and (max-width: 1200px)';
